LayoutTests/imported/w3c/web-platform-tests/websockets/Send-unicode-data.any.worker-expected.txt


FAIL Send unicode data on a WebSocket - Connection should be closed assert_equals: expected 0 but got 12
